pyproject.toml
[tool.poetry]
name = "translation_with_synchronous_grammar"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
torch = {version = "^2.3.1+cu121", source = "torch_cu121"}
nltk = "^3.6.5"
subword-nmt = "^0.3.7"
sacremoses = "^0.0.47"
stanza = "^1.3.0"
fairseq = "^0.12.2"

[tool.poetry.dev-dependencies]
black = "^21.11b1"
isort = "^5.10.1"
mypy = "^0.910"
flake8 = "^4.0.1"
absl-py = "^1.0.0"
compare-mt = "^0.2.8"

[[tool.poetry.source]]
name = "torch_cu121"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
